renames fetch-secret-config decorator
