+++ ui/lib/kubernetes/addon/routes/configuration.js
@@ -5,7 +5,7 @@
 
 import Route from '@ember/routing/route';
 import { inject as service } from '@ember/service';
-import { withConfig } from 'core/decorators/fetch-secret-config';
+import { withConfig } from 'core/decorators/fetch-secrets-engine-config';
 
 @withConfig('kubernetes/config')
 export default class KubernetesConfigureRoute extends Route {
